Deduplicate TR2 logic
