+++ TRRandomizerCore/Randomizers/TR2/Classic/TR2AudioRandomizer.cs
@@ -1,10 +1,4 @@
-﻿using Newtonsoft.Json;
-using TRGE.Core;
-using TRLevelControl;
-using TRLevelControl.Helpers;
-using TRLevelControl.Model;
-using TRRandomizerCore.Levels;
-using TRRandomizerCore.SFX;
+﻿using TRGE.Core;
 
 namespace TRRandomizerCore.Randomizers;
 
@@ -12,224 +6,27 @@ public class TR2AudioRandomizer : BaseTR2Randomizer
 {
     private const int _numSamples = 408; // Number of entries in MAIN.SFX
 
-    private AudioRandomizer _audioRandomizer;
-
-    private List<TR2SFXDefinition> _soundEffects;
-    private List<TRSFXGeneralCategory> _sfxCategories;
-
     public override void Randomize(int seed)
     {
-        _generator = new(seed);
-        LoadAudioData();
+        TR2AudioAllocator allocator = new(ScriptEditor.AudioProvider.GetCategorisedTracks(), _numSamples)
+        {
+            Generator = new(seed),
+            Settings = Settings,
+        };
+        allocator.Initialise(Levels.Select(l => l.LevelFileBaseName), BackupPath);
 
         foreach (TR2ScriptedLevel lvl in Levels)
         {
             LoadLevelInstance(lvl);
 
-            RandomizeMusicTriggers(_levelInstance);
-            RandomizeSoundEffects(_levelInstance);
-            RandomizeWibble(_levelInstance);
+            allocator.RandomizeMusicTriggers(_levelInstance.Data);
+            allocator.RandomizeSoundEffects(_levelInstance.Name, _levelInstance.Data);
+            allocator.RandomizePitch(_levelInstance.Data);
 
             SaveLevelInstance();
             if (!TriggerProgress())
             {
                 break;
-            }
-        }
-    }
-
-    private void RandomizeMusicTriggers(TR2CombinedLevel level)
-    {
-        if (Settings.ChangeTriggerTracks)
-        {
-            RandomizeFloorTracks(level.Data);
-        }
-
-        if (Settings.SeparateSecretTracks)
-        {
-            RandomizeSecretTracks(level.Data);
-        }
-    }
-
-    private void RandomizeFloorTracks(TR2Level level)
-    {
-        _audioRandomizer.ResetFloorMap();
-        foreach (TR2Room room in level.Rooms)
-        {
-            _audioRandomizer.RandomizeFloorTracks(room, level.FloorData);
-        }
-    }
-
-    private void RandomizeSecretTracks(TR2Level level)
-    {
-        List<TRAudioTrack> secretTracks = _audioRandomizer.GetTracks(TRAudioCategory.Secret);
-        Dictionary<int, TR2Entity> secrets = GetSecretItems(level);
-        foreach (int entityIndex in secrets.Keys)
-        {
-            TR2Entity secret = secrets[entityIndex];
-            TRRoomSector sector = level.GetRoomSector(secret);
-            if (sector.FDIndex == 0)
-            {
-                level.FloorData.CreateFloorData(sector);
-            }
-
-            List<FDEntry> entries = level.FloorData[sector.FDIndex];
-            FDTriggerEntry existingTriggerEntry = entries.Find(e => e is FDTriggerEntry) as FDTriggerEntry;
-            bool existingEntityPickup = false;
-            if (existingTriggerEntry != null)
-            {
-                if (existingTriggerEntry.TrigType == FDTrigType.Pickup && existingTriggerEntry.Actions[0].Parameter == entityIndex)
-                {
-                    // GW gold secret (default location) already has a pickup trigger to spawn the
-                    // TRex (or whatever enemy) so we'll just append to that item list here
-                    existingEntityPickup = true;
-                }
-                else
-                {
-                    // There is already a non-pickup trigger for this sector so while nothing is wrong with
-                    // adding a pickup trigger, the game ignores it. So in this instance, the sound that
-                    // plays will just be whatever is set in the script.
-                    continue;
-                }
-            }
-
-            FDActionItem musicAction = new()
-            {
-                Action = FDTrigAction.PlaySoundtrack,
-                Parameter = (short)secretTracks[_generator.Next(0, secretTracks.Count)].ID
-            };
-
-            if (existingEntityPickup)
-            {
-                existingTriggerEntry.Actions.Add(musicAction);
-            }
-            else
-            {
-                entries.Add(new FDTriggerEntry
-                {
-                    TrigType = FDTrigType.Pickup,
-                    Actions = new()
-                    {
-                        new() { Parameter = (short)entityIndex },
-                        musicAction
-                    }
-                });
-            }
-        }
-    }
-
-    private static Dictionary<int, TR2Entity> GetSecretItems(TR2Level level)
-    {
-        Dictionary<int, TR2Entity> entities = new();
-        for (int i = 0; i < level.Entities.Count; i++)
-        {
-            if (TR2TypeUtilities.IsSecretType(level.Entities[i].TypeID))
-            {
-                entities[i] = level.Entities[i];
-            }
-        }
-
-        return entities;
-    }
-
-    private void LoadAudioData()
-    {
-        _audioRandomizer = new(ScriptEditor.AudioProvider.GetCategorisedTracks())
-        {
-            Generator = _generator,
-            Settings = Settings,
-        };
-        _audioRandomizer.ChooseUncontrolledLevels(new(Levels.Select(l => l.LevelFileBaseName)), TR2LevelNames.ASSAULT);
-
-        _sfxCategories = AudioRandomizer.GetSFXCategories(Settings);
-        if (_sfxCategories.Count > 0)
-        {
-            _soundEffects = JsonConvert.DeserializeObject<List<TR2SFXDefinition>>(ReadResource(@"TR2\Audio\sfx.json"));
-
-            Dictionary<string, TR2Level> levels = new();
-            TR2LevelControl reader = new();
-            foreach (TR2SFXDefinition definition in _soundEffects)
-            {
-                if (!levels.ContainsKey(definition.SourceLevel))
-                {
-                    levels[definition.SourceLevel] = reader.Read(Path.Combine(BackupPath, definition.SourceLevel));
-                }
-
-                TR2Level level = levels[definition.SourceLevel];
-                definition.SoundEffect = level.SoundEffects[definition.InternalIndex];
-            }
-        }
-    }
-
-    private void RandomizeSoundEffects(TR2CombinedLevel level)
-    {
-        if (_sfxCategories.Count == 0)
-        {
-            return;
-        }
-
-        if (_audioRandomizer.IsUncontrolledLevel(level.Name))
-        {
-            // Choose a random but unique pointer into MAIN.SFX for each sample.
-            HashSet<uint> indices = new();
-            foreach (TR2SoundEffect effect in level.Data.SoundEffects.Values)
-            {
-                do
-                {
-                    effect.SampleID = (uint)_generator.Next(0, _numSamples + 1 - Math.Max(effect.SampleCount, 1));
-                }
-                while (!indices.Add(effect.SampleID));
-            }
-        }
-        else
-        {
-            foreach (TR2SFX internalIndex in Enum.GetValues<TR2SFX>())
-            {
-                TR2SFXDefinition definition = _soundEffects.Find(sfx => sfx.InternalIndex == internalIndex);
-                if (!level.Data.SoundEffects.ContainsKey(internalIndex) || definition == null
-                    || definition.Creature == TRSFXCreatureCategory.Lara || !_sfxCategories.Contains(definition.PrimaryCategory))
-                {
-                    continue;
-                }
-
-                Predicate<TR2SFXDefinition> pred;
-                if (Settings.LinkCreatureSFX && definition.Creature > TRSFXCreatureCategory.Lara)
-                {
-                    pred = sfx =>
-                    {
-                        return sfx.Categories.Contains(definition.PrimaryCategory) &&
-                        sfx != definition &&
-                        (
-                            sfx.Creature == definition.Creature ||
-                            (sfx.Creature == TRSFXCreatureCategory.Lara && definition.Creature == TRSFXCreatureCategory.Human)
-                        );
-                    };
-                }
-                else
-                {
-                    pred = sfx => sfx.Categories.Contains(definition.PrimaryCategory) && sfx != definition;
-                }
-
-                List<TR2SFXDefinition> otherDefinitions = _soundEffects.FindAll(pred);
-                if (otherDefinitions.Count > 0)
-                {
-                    TR2SFXDefinition nextDefinition = otherDefinitions[_generator.Next(0, otherDefinitions.Count)];
-                    if (nextDefinition != definition)
-                    {
-                        level.Data.SoundEffects[internalIndex] = nextDefinition.SoundEffect;
-                    }
-                }
-            }
-        }
-    }
-
-    private void RandomizeWibble(TR2CombinedLevel level)
-    {
-        if (Settings.RandomizeWibble)
-        {
-            foreach (var (_, effect) in level.Data.SoundEffects)
-            {
-                effect.RandomizePitch = true;
             }
         }
     }
